# lib/api-zod/src/generated/types/ambiguityResolution.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Scopefold provenance-first operations API
 * OpenAPI spec version: 0.1.0
 */

export interface AmbiguityResolution {
  ambiguityId: string;
  optionId: string;
}