# lib/api-zod/src/generated/types/bootstrap.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Scopefold provenance-first operations API
 * OpenAPI spec version: 0.1.0
 */
import type { BootstrapWorkspace } from './bootstrapWorkspace';
import type { BrandSettings } from './brandSettings';
import type { Project } from './project';

export interface Bootstrap {
  workspace: BootstrapWorkspace;
  projects: Project[];
  settings: BrandSettings;
}