# lib/api-zod/src/generated/types/brandSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Scopefold provenance-first operations API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandSettings {
  agencyName: string;
  brandName: string;
  primaryColor: string;
  evidenceColor: string;
  currency: string;
  clientTerm: string;
  projectTerm: string;
  milestoneTerm: string;
}